Implemented STA

+++ CPU.h
@@ -116,7 +116,7 @@ class CPU {
     void SEC();
     void SED();
     void SEI();
-    void STA(uint8_t);
+    void STA(uint16_t);
     void STX(uint8_t);
     void STY(uint8_t);
     void TAX();
